#1212 share test model functions between unit/instrumentation tests
