+++ app/build.gradle
@@ -258,6 +258,16 @@ android {
       }
     }
   }
+  sourceSets {
+    test {
+      java.srcDirs += "$projectDir/src/testShared"
+    }
+
+    androidTest {
+      java.srcDirs += "$projectDir/src/testShared"
+    }
+  }
+
 
   flavorDimensions "default"
